+++ modules/katamari/src/main/ts/ephox/katamari/api/Obj.ts
@@ -3,6 +3,11 @@ import { Eq } from '@ephox/dispute';
 import * as Fun from './Fun';
 import { Optional } from './Optional';
 
+type ObjKeys<T extends {}> = Extract<keyof T, string>;
+type ObjCallback<T extends {}> = (value: T[keyof T], key: ObjKeys<T>) => void;
+type ObjMorphism<T extends {}, R> = (value: T[keyof T], key: ObjKeys<T>) => R;
+type ObjPredicate<T extends {}> = (value: T[keyof T], key: ObjKeys<T>) => boolean;
+
 // There are many variations of Object iteration that are faster than the 'for-in' style:
 // http://jsperf.com/object-keys-iteration/107
 //
@@ -12,8 +17,8 @@ export const keys = Object.keys;
 // eslint-disable-next-line @typescript-eslint/unbound-method
 export const hasOwnProperty = Object.hasOwnProperty;
 
-export const each = <T>(obj: T, f: (value: T[keyof T], key: string) => void): void => {
-  const props = keys(obj);
+export const each = <T extends {}>(obj: T, f: ObjCallback<T>): void => {
+  const props = keys(obj) as Array<ObjKeys<T>>;
   for (let k = 0, len = props.length; k < len; k++) {
     const i = props[k];
     const x = obj[i];
@@ -21,14 +26,14 @@ export const each = <T>(obj: T, f: (value: T[keyof T], key: string) => void): vo
   }
 };
 
-export const map = <T, R>(obj: T, f: (value: T[keyof T], key: string) => R): {[k in keyof T]: R} => {
+export const map = <T extends {}, R>(obj: T, f: ObjMorphism<T, R>): {[k in keyof T]: R} => {
   return tupleMap<{[k in keyof T]: R}, T>(obj, (x, i) => ({
     k: i,
     v: f(x, i)
   }));
 };
 
-export const tupleMap = <R, T>(obj: T, f: (value: T[keyof T], key: string) => {k: string; v: any}): R => {
+export const tupleMap = <R extends {}, T extends {}>(obj: T, f: ObjMorphism<T, { k: keyof R; v: R[keyof R] }>): R => {
   const r = {} as R;
   each(obj, (x, i) => {
     const tuple = f(x, i);
@@ -37,32 +42,30 @@ export const tupleMap = <R, T>(obj: T, f: (value: T[keyof T], key: string) => {k
   return r;
 };
 
-const objAcc = <K extends number | string | symbol, V>(r: Record<K, V>) => (x: V, i: K): void => {
+const objAcc = <T extends {}>(r: T) => (x: T[keyof T], i: keyof T): void => {
   r[i] = x;
 };
 
-const internalFilter = <V>(obj: Record<string, V>, pred: (value: V, key: string) => boolean, onTrue: (value: V, key: string) => void, onFalse: (value: V, key: string) => void) => {
-  const r: Record<string, V> = {};
+const internalFilter = <T extends {}>(obj: T, pred: ObjPredicate<T>, onTrue: ObjCallback<T>, onFalse: ObjCallback<T>) => {
   each(obj, (x, i) => {
     (pred(x, i) ? onTrue : onFalse)(x, i);
   });
-  return r;
 };
 
-export const bifilter = <V>(obj: Record<string, V>, pred: (value: V, key: string) => boolean): {t: Record<string, V>; f: Record<string, V>} => {
-  const t: Record<string, V> = {};
-  const f: Record<string, V> = {};
+export const bifilter = <T extends {}>(obj: T, pred: ObjPredicate<T>): { t: Record<string, T[keyof T]>; f: Record<string, T[keyof T]> } => {
+  const t: Record<string, T[keyof T]> = {} ;
+  const f: Record<string, T[keyof T]> = {};
   internalFilter(obj, pred, objAcc(t), objAcc(f));
   return { t, f };
 };
 
-export const filter = <V>(obj: Record<string, V>, pred: (value: V, key: string) => boolean): Record<string, V> => {
-  const t: Record<string, V> = {};
+export const filter = <T extends {}>(obj: T, pred: ObjPredicate<T>): Record<string, T[keyof T]> => {
+  const t: Record<string, T[keyof T]> = {};
   internalFilter(obj, pred, objAcc(t), Fun.noop);
   return t;
 };
 
-export const mapToArray = <T, R>(obj: T, f: (value: T[keyof T], key: string) => R): R[] => {
+export const mapToArray = <T extends {}, R>(obj: T, f: ObjMorphism<T, R>): R[] => {
   const r: R[] = [];
   each(obj, (value, name) => {
     r.push(f(value, name));
@@ -70,8 +73,8 @@ export const mapToArray = <T, R>(obj: T, f: (value: T[keyof T], key: string) => 
   return r;
 };
 
-export const find = <T>(obj: T, pred: (value: T[keyof T], key: string, obj: T) => boolean): Optional<T[keyof T]> => {
-  const props = keys(obj);
+export const find = <T extends {}>(obj: T, pred: (value: T[keyof T], key: ObjKeys<T>, obj: T) => boolean): Optional<T[keyof T]> => {
+  const props = keys(obj) as Array<ObjKeys<T>>;
   for (let k = 0, len = props.length; k < len; k++) {
     const i = props[k];
     const x = obj[i];
@@ -82,7 +85,7 @@ export const find = <T>(obj: T, pred: (value: T[keyof T], key: string, obj: T) =
   return Optional.none();
 };
 
-export const values = <T>(obj: T): Array<T[keyof T]> => {
+export const values = <T extends {}>(obj: T): Array<T[keyof T]> => {
   return mapToArray(obj, Fun.identity);
 };
 
@@ -90,14 +93,14 @@ export const size = (obj: {}): number => {
   return keys(obj).length;
 };
 
-export const get = <T, K extends keyof T>(obj: T, key: K): Optional<NonNullable<T[K]>> => {
+export const get = <T extends {}, K extends keyof T>(obj: T, key: K): Optional<NonNullable<T[K]>> => {
   return has(obj, key) ? Optional.from(obj[key] as NonNullable<T[K]>) : Optional.none();
 };
 
-export const has = <T, K extends keyof T>(obj: T, key: K): boolean =>
+export const has = <T extends {}, K extends keyof T>(obj: T, key: K): boolean =>
   hasOwnProperty.call(obj, key);
 
-export const hasNonNullableKey = <T, K extends keyof T>(obj: T, key: K): obj is T & Record<K, NonNullable<T[K]>> =>
+export const hasNonNullableKey = <T extends {}, K extends keyof T>(obj: T, key: K): obj is T & Record<K, NonNullable<T[K]>> =>
   has(obj, key) && obj[key] !== undefined && obj[key] !== null;
 
 export const isEmpty = (r: Record<any, any>): boolean => {