+++ modules/katamari/src/main/ts/ephox/katamari/api/Unique.ts
@@ -2,7 +2,7 @@ import * as Arr from './Arr';
 import * as Obj from './Obj';
 
 export const stringArray = (a: string[]): string[] => {
-  const all = {};
+  const all: Record<string, {}> = {};
   Arr.each(a, (key) => {
     all[key] = {};
   });